Added TOP_FILES and removed static README and LICENSE file
TopFiles are what are used to be README and LICENSE,
they are configurable in config.h(at compile-time) but are checked for existense at runtime(like old README and LICENSE).
It can be used to add more files on the top bar(for example Makefile or in general often present files)
I was thinking abount a runtime(argv) selectable files but it is not implemented.